job_cost_optimization_demo/requirements.txt
# Databricks Job 비용 최적화 데모 - 필수 패키지

# Databricks SDK (필수)
databricks-sdk>=0.18.0

# 환경변수 관리
python-dotenv>=1.0.0

# Slack 알림 (옵션)
requests>=2.31.0

# 재시도 로직 (옵션)
tenacity>=8.2.0

# AWS Secrets Manager 통합 (옵션)
# boto3>=1.34.0

# 로깅 및 모니터링 (옵션)
# colorlog>=6.8.0
